src/radar_data/infrastructure/config/loader.py
"""Configuration loader for YAML files.

This module loads and validates configuration files (datasets.yml, quality.yml)
using Pydantic models defined in schema.py.
"""

from pathlib import Path
from typing import Any

import yaml

from radar_data.infrastructure.config.schema import DatasetsConfig, QualityProfile


def load_configs(
    datasets_path: str | Path,
    quality_path: str | Path | None = None,
) -> tuple[list[dict[str, Any]], dict[str, dict[str, Any]]]:
    """Load dataset and quality configuration files.

    Args:
        datasets_path: Path to datasets.yml file.
        quality_path: Optional path to quality.yml file.
                     If None, returns empty quality profiles.

    Returns:
        Tuple of (datasets_list, quality_profiles_dict).
        - datasets_list: List of dataset configuration dictionaries.
        - quality_profiles_dict: Dictionary mapping profile names to configs.

    Raises:
        FileNotFoundError: If config files do not exist.
        yaml.YAMLError: If YAML files are malformed.
        ValueError: If configuration does not match schema.

    TODO:
        - Implement actual YAML loading and validation.
        - Use Pydantic models from schema.py for validation.
        - Add caching of loaded configs.
        - Add support for environment variable substitution.
    """
    # TODO: Implement actual loading
    # datasets_file = Path(datasets_path)
    # if not datasets_file.exists():
    #     raise FileNotFoundError(f"Datasets config not found: {datasets_path}")
    #
    # with open(datasets_file) as f:
    #     data = yaml.safe_load(f)
    #     config = DatasetsConfig(**data)
    #     datasets = [ds.dict() for ds in config.datasets]
    #
    # quality_profiles = {}
    # if quality_path:
    #     quality_file = Path(quality_path)
    #     if quality_file.exists():
    #         with open(quality_file) as f:
    #             data = yaml.safe_load(f)
    #             profile_config = QualityProfile(**data)
    #             quality_profiles = {
    #                 name: cfg.dict() for name, cfg in profile_config.profiles.items()
    #             }
    #
    # return datasets, quality_profiles
    pass

