src/radar_data/infrastructure/normalize/mapping.py
"""External to internal mapping adapter implementing NormalizerPort."""

from datetime import datetime
from typing import Any

from dateutil import parser as date_parser

from radar_data.domain.entities import Observation, Provenance, Series
from radar_data.domain.interfaces import NormalizerPort


class MappingNormalizer(NormalizerPort):
    """Adapter for normalizing raw parsed data into domain entities.

    Implements the NormalizerPort interface for mapping external schemas
    to internal domain entities. Handles column mapping, date parsing,
    and entity creation.

    TODO:
        - Implement actual normalization logic.
        - Add column mapping from config.
        - Add date parsing with error handling.
        - Add unit conversion support.
        - Add Series and Observation creation.
        - Add provenance tracking.
    """

    def normalize(
        self,
        raw_records: list[dict[str, str | float]],
        dataset_config: dict[str, Any],
    ) -> tuple[Series, list[Observation]]:
        """Normalize raw records into domain entities.

        Args:
            raw_records: List of raw dictionaries from parser.
            dataset_config: Configuration dict with normalize section.

        Returns:
            Tuple of (Series, list[Observations]).

        Raises:
            ValueError: If normalization fails (e.g., missing required fields).

        TODO:
            - Extract normalize config (internal_series_code, unit, frequency).
            - Create Series entity.
            - Map raw records to Observations:
              - Parse dates using dateutil.
              - Extract values (handle string/number conversion).
              - Create Observation entities.
            - Add provenance metadata if available.
        """
        # TODO: Implement normalization
        # normalize_config = dataset_config.get("normalize", {})
        # source_config = dataset_config.get("source", {})
        #
        # # Create Series
        # series = Series(
        #     code=normalize_config.get("internal_series_code"),
        #     name=dataset_config.get("name", ""),
        #     unit=normalize_config.get("unit", ""),
        #     frequency=normalize_config.get("frequency", ""),
        #     provider=dataset_config.get("provider", ""),
        # )
        #
        # # Map raw records to Observations
        # observations = []
        # columns = source_config.get("columns", {})
        # date_col = columns.get("date", "date")
        # value_col = columns.get("value", "value")
        #
        # for record in raw_records:
        #     date_str = str(record.get(date_col, ""))
        #     value_str = str(record.get(value_col, ""))
        #
        #     # Parse date
        #     try:
        #         timestamp = date_parser.parse(date_str)
        #     except (ValueError, TypeError) as e:
        #         raise ValueError(f"Failed to parse date: {date_str}") from e
        #
        #     # Parse value
        #     try:
        #         value = float(value_str)
        #     except (ValueError, TypeError) as e:
        #         raise ValueError(f"Failed to parse value: {value_str}") from e
        #
        #     # Create Observation
        #     obs = Observation(
        #         series_code=series.code,
        #         timestamp=timestamp,
        #         value=value,
        #     )
        #     observations.append(obs)
        #
        # return series, observations
        pass

