pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "radar-data-pipeline"
version = "0.1.0"
description = "Data ingestion, normalization, and cleaning service"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["data-engineering", "etl", "data-pipeline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
]

[project.scripts]
radar-data = "radar_data.interface.cli.runner:main"

[tool.setuptools]
packages = ["radar_data"]

[tool.setuptools.package-data]
radar_data = ["py.typed"]

